backend/requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.27.1
PyPDF2==3.0.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
pydantic==2.6.3